requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
pandas>=2.2.0
streamlit>=1.36.0
pillow>=10.4.0
requests>=2.32.0
mediapipe>=0.10.0
opencv-python>=4.10.0
numpy>=1.26.0
sentence-transformers>=3.1.0




